backend/requirements.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
Flask-Cors==3.0.9
ImageHash==4.1.0
imageio==2.9.0
imgaug==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
rank-bm25==0.2.1
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
tifffile==2020.10.1
Werkzeug==1.0.1
